Add "foreground" mode for docker run This is a mode of "docker run" that uses direct forking from the docker client, meaning that the container is a direct child of the docker run command (rather than the daemon), and that it directly inherits stdout/err (and stdin if requested) from the parent of the "docker run" command.
This is very useful in cases where you e.g. want to run a docker
container from systemd (or other process manager), as we can use the
process manager to e.g. set up cgroups and limits, monitor and
restart, pass in fds for socket activation, etc.

The way this works is that there is a new execdriver, which is used
for forground containers. This container just does rpc calls back to
the "docker run" command, which in turn uses a "real" execdriver and
feeds it whatever data the "fake" execdriver got, but in the context
of the docker run command.

We feed information about the process (exit status, pid, etc) to the
daemon, so it is visible just like any other docker container via the
docker apis.